requirements.txt
pandas
mysql-connector-python
binance
pytz
logging 